Fix an issue where bubble bar would collapse when rotating
When the device is rotated, everything gets recreated, in this case,
bubble bar state is typically set to be stashed. However, if taskbar
and bubble bar were unstashed while rotating, bubble bar would be
incorrectly stashed. To fix this, only do the stash if taskbar is
also stashed in this case.

Flag: com.android.wm.shell.enable_bubble_bar
Test: manual - be in an app, unstash bubbles, rotate device
             => observe that bubble bar is unstashed like taskbar
                after rotation
Bug: 353884121
Change-Id: Ia37397c62a1a6819c1bd42261f8a9955db511da3