Update stashing hotseat logic.
Before this change we only aligned the taskbar with the hotseat if the
hotseat position is changed. But it does not work if hoseat icons have
reduced space in between each other to accommodate the bubble bar. This
change aligns the task bar with the hotseat icons before any hotseat
animation.

Fixes: 372075963
Flag: com.android.wm.shell.enable_bubble_bar
Test: Manual. Have the hotseat with the reduced icons space. Expand /
collapse the bubble bar.

Change-Id: I72156cfc78f4db93b5bba11f1bfdda83c20b9a02