Fixed the opened folder on taskbar when NotifShade is shown
This cl fixes the bug by closing all abstract floating views
including opened folders when the notification shade is shown.
Usually this was handled in the folder class where it intercepted
the touch event to close it by dispatching the back key, but this
didn't work when the virtual keyboard existed.

More context is added in the bug.

Flag: EXEMPT bug fix
Fixes: 267529599
Test: Ran on a 1P device

Change-Id: Ifdd357668ece2c11888ecfc69af64b1742c07d6e