Fix split selection translation when focused task is selected for splitscreen
After enabling large tile for desktop windowing, some animations and translations for dismiss were changed. The changes involved to translate the tasks when a large tile (i.e., focused task and desktop window) are dismissed and no other focusedTaskView is selected to replace the previous one. However, when we split a focused task, we don't replace it by another task in the grid. And also triggers the createTaskDismissAnimation, entering the same condition described before (focused task dismissed, nextFocusedTaskView = null), causing the weird and broken animation for split.

To fix this, we skip this animation when is dismissing a task for split selection.

Fix: 374104009
Flag: com.android.launcher3.enable_large_desktop_windowing_tile
Test: Manual. With flag enabled and multiple apps opened in Overview, split the focused task.
Change-Id: Ib79ad5abc0660a0b654f23b22b90f9889c6f7806