Fix Taskbar Touchable Region when overview is in split select mode
The problem: when Overview is in split slect mode second split app pick from taskbar take you back home. Why, because upon touch the taskbar touch region get
 set to empty and then we later don't meet any conditions to set default touch region to something.

The solution: change the condition of checking overview in onTaskbarOrBubblebarWindowHeightOrInsetsChanged so that we check that taskbar is in any overview s
tate and we don't rely on the stash controller flag alone. we follow similar logic in updateInsetsTouchability, but we gate that only behind transient taskba
r hence this was not an issue for transient taskbar.

Test: Manual
Bug: 373559050
Flag: EXEMPT bugfix
Change-Id: Ia507354f42959ce9d320114a313bfc533186b524