Don't allow Desktop tasks to go outside Overview task bounds
Wrap TaskThumbnailView's in a new DesktopTaskContentView. This is in order to correctly crop the thumbnail children contained in the DesktopTaskView, otherwise any freeform task screenshots at the edges of the DTV leak outside the bounds.

We need a new container to do this, otherwise we will cut off the icon
or the hover border if we apply an outline to the DesktopTaskView itself.

Fix: 376206104
Flag: com.android.window.flags.enable_desktop_windowing_mode
Test: OverviewDesktopTaskImageTest
Change-Id: Ic51438e63aaf671a0a8156815d11fc7fa12beac0