Tuning gesture nav params for trackpad 3-finger gestures
It's harder to slow down in trackpad swipes than finger swipes on the screen, so the deceleration required to be considered pause for swiping from app should be less strict (set the number for trackpad to 0.85f instead of 0.6f for on-screen). Also added a speed_trackpad_somewhat_fast for the case of swiping up to overview, so that we can be a little bit more lenient to the users for the end speed.

Bug: 375688524
Test: play with 3-finger swipe to home / overview. Make it easier to go to overview, but fast swipe from app still takes user to home
Flag: EXEMPT bugfix
Change-Id: I196a92163fd36e9b7013f4893876b79c9c2c9de0