Start background tasks via task instead of intent.
When the second app is selected, check if there is a background task
that matches the intent provided and start the task if there
is one.

This is done to correct an issue where freeform background tasks opened
via intent will open in the correct split stage but in freeform
windowing mode and bounds.

Bug: 338382550
Test: manual
Flag: EXEMPT bugfix
Change-Id: Ibefe4b5045cbcefdbc71c99c4eaba4fdf67d7a0e