Update the flyout collapsed position
This change ensures that we're updating the distance to the flyout's
collapsed position before we start collapsing it. The distance is
initially calculated before we start expanding, but when a new bubble
is added while we're fully expanded, and if we previously had 1
bubble and now have 2, the collapsed position changed, so we need to
update it before we start collapsing.

Flag: com.android.wm.shell.enable_bubble_bar
Bug: 376276297
Test: manual
      - dismiss all bubbles
      - create 1 bubble
      - after the flyout is fully expanded create a second bubble
      - observe the flyout is updated with the contents of the second
        bubble
      - observe when the flyout collapses, it collapses to the dot
        on the second bubble
Change-Id: I219886fd6463d69692d9945d901d1f3ad3cb26a0