Don't crop bubble flyout in overview
When switching to overview we set the taskbar window size as part
of the taskbar animation. If a bubble is animating the window size
becomes too small and the bubble flyout gets cropped. This change
avoids updating the window size in the middle of a bubble animation.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 377809168
Test: manual
       - create a bubble and wait for the flyout to show
       - switch to overview
       - observe the flyout is not cropped
Change-Id: Ic0a3f2c0fc534162ae5583e744a7bc400c3d921b