Draw contrast tile around the text
If the BubbletTextView should have an app contrast tile, account for
some padding between the tile and the text when measuring the view, so
that the text is guaranteed to be inside of the bubble.  In a similar
way, recalculate the bounds of the tile to only surround the app title
as needed.

Bug: 341217082
Flag: com.android.launcher3.enable_contrast_tiles
Test: Manual, visual change
Change-Id: I9e8b149ade3fa4522e62b89c2332195f648b190f