Fix bubble animation when swiping home
When swiping home during the bubble animation, and if the show edu
view state field is on, we try to show the bubble bar as part of
preparing to show the edu view. This updates the stash state and
prevents the animation from getting interrupted.

The expected behavior around showing the edu view is to store the
request when it is received and show the view later after the user
taps on the bubble bar or the flyout to expand it. We shouldn't
attempt to show the bubble bar at that point.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 378967270
Bug: 374842575
Test: manual
    - force show the edu: adb shell settings put secure force_show_bubbles_user_education 1
    - launch an app and send a bubble
    - while the bubble animates, swipe to go home
    - observe the flyout is removed correctly, and we don't end up in stashed state
Change-Id: I478b06ac5f6f9d398cd143bcc0b3f796d9ad8a49