Show the dot after the flyout is interrupted
When the bubble notification animation is interrupted, we now
animate the dot back in so that it is visible.

Before this change we unsuppressed the dot which should normally
make the dot visible. The problem was that when the animation is
interrupted we immediately clear the animating bubble field but
showing the dot happens at the end of the flyout fade animation.

This change also animates the dot back in instead of setting the
scale immediately to 1.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 379733568
Test: manual
       - launch app
       - send a bubble
       - when the flyout shows, swipe to go home
       - observe animation is cleared and dot is visible
Change-Id: Ie7a10ddf237480ac6555d5f2b40da343394ce77f