Fix bubble bar stash delay
Bubble bar was stashing with a delay compared to taskbar.
This is because taskbar stashing looks at isInLauncher value in addition
to launcher state.
Add this check to the bubble bar in overview condition. And only
consider bubble bar showing in overview if we isInLauncher returns true
and launcher state is overview.

Bug: 371229068
Test: manual, start from home with bubble bar shown, swipe up to
  overview, tap on an app, observe that bubble bar stashes at the same
  time as taskbar
Test: swipe up to home, open all apps, check that bubble bar remains
  visible
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: Icedfd29300f739e37d636122bad0fb6e14fa2ced