Revert "Revert "Use the Coreographer's frame time for a more rel..."
Revert submission 30577378-revert-30509816-raf-timing-EAFMYZVSQM

Reason for revert: this test on Cuttlefish seems to be very unstable in general. I tried to re-test on ABTD with these reverts patched in, but the test isn't being included in the suite, and I couldn't find information anywhere on whether it's supposed to be still running. Since these changes are fully flagged, I don't see how they could be the real culprit. Trying to submit again and monitoring the impact on the postsubmits.

Reverted changes: /q/submissionid:30577378-revert-30509816-raf-timing-EAFMYZVSQM

Change-Id: I4c1dc43858060716450bab20e3c56728235508c7