Fix handle color when interrupting the animation
Pass the correct stash value to the controller after interrupting
an ongoing animation due to stash state changing. We previously
incorrectly passed a value that depended on the alpha value of the
bubble bar. This allows sampling to happen and for the handle to
have the correct color.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 377604151
Test: atest BubbleBarViewAnimatorTest
Test: manual
       1. Create 1+ bubble
       2. Receive new message and see the flyout
       3. Open an app immediately (Don't wait until the flyout disappear)
Change-Id: I739bcbd8fe4f7e02285244d78d7159c571be1ec2