Prevent bad grid configuration by making sure RestoreDbTask updates IDP correctly.
IPD#reset() was resetting the internal IDP dbFile to the factory default instead of current grid/db

This lead to LoaderTask and grid migration creating an empty DB for the wrong grid size, while the restored DB remained unused. Leading to an empty home screen because an empty DB file was used.

Bug: 377510898
Flag: EXEMPT bugfix
Test: manually testing backup & restore with various grid sizes
Change-Id: I709201e484afb8f12a4fde21198d2c5793d89ca1