Fix bar visibility after dismissing while stashed
When updating the bubble bar visibility, allow it to become visible
if the only hidden bit is the stashed one and we're animating.

This fixes the visibility issue that appears after dismissing the
bar while stashed, and then sending an initial bubble, while still
stashed.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 383620113
Test: manual
       - create bubble
       - launch app
       - dismiss bubble with adb
       - create bubble with adb
       - observe bubble bar is visible during the animation
Change-Id: I7a955de068867829b3ccab11b2b0adc8d8e10239