Don't update desktop visibility with final state until state transition has finished.
When exiting Overview, the Launcher.onIdpChange listener is indirectly
triggered by DesktopVisibilityController.onLauncherStateChanged because
with a desktop live tile the mode will have changed out of desktop mode.

onIdpChange is treated as a configuration change and triggers reapplyUi
which cancels currently running animations and causes the Overview->Home
animation not to run.

Bug: 378481525
Test: locally tested on TangorPro
Flag: EXEMPT bugfix
Change-Id: I1629e43d7b2fd960c5b999899b286e918a1e7aa4