chore(magnification): remove the PRIVATE_FLAG_EXCLUDE_FROM_SCREEN_MAGNIFICATION window flag usage in Launcher app
As the bug, in some case the launcher taskbar is visible but the taskbar
window is flagged with PRIVATE_FLAG_EXCLUDE_FROM_SCREEN_MAGNIFICATION, so
the magnification consumes the gesture events on the taskabr, then the
accessibility gesture shortcut cannot take effect.

The flag was introduced in Launcher to fix the fullscreen magnification border
cutting/overlapping with taskbar issue. Since we already migrate the
fullscreen border to sysui and make it always fit the screen, regardless
the magnification regions, so we can remove the flag usage in Launcher
to solve the bug and do not affect the fullscreen border layout correctness.

We'll use a bugfix feature flag to gradually rollout this change.

To test the fix, on LargeScreen devices, we can enable accessibility gesture
shortcut and disable the "keep on while switching apps" toggle in
Magnification Settings page, then in any app we use the 2 finger swipe gestures
to try activate/deactivate the magnification. When the flag is off, swipe
gesture can not successfully deactivate the magnification, while the
flah is on it work. (attached records in bug)

Bug: 369019568
Flag: com.android.launcher3.remove_exclude_from_screen_magnification_flag_usage
Test: manually (see message above)
Change-Id: I469189a2288d0ddbfb5a5d9e02a78366870ce5b1